correct: divideNumberInCountOfDivdideLevels function in util
