+++ src/utils/utils.js
@@ -3,3 +3,27 @@ export function entryObject(o) {
 
   return Object.entries(o);
 }
+
+export function divideNumberInCountOfDivideLevels(totalAmount, divideLevels) {
+  if (divideLevels.some((divideLevel) => typeof divideLevel !== 'number')) {
+    throw new Error('divideLevels should be consisted in numbers');
+  }
+
+  if (divideLevels.some((divideLevel, i) => {
+    if (i > 0) {
+      const prevDivideLevel = divideLevels[i - 1];
+      return prevDivideLevel <= divideLevel;
+    }
+    return false;
+  })) {
+    throw new Error('divideLevels should be sorted in descending order');
+  }
+
+  let currentTotalAmount = totalAmount;
+  return divideLevels.map((divideLevel, i) => {
+    const count = Math.floor(currentTotalAmount / divideLevel);
+    currentTotalAmount -= count * divideLevel;
+
+    return count;
+  });
+}
